+++ CHANGES.md
@@ -4,6 +4,8 @@ Release Notes.
 
 8.16.0
 ------------------
+
+* Exclude `synthetic` methods for the WitnessMethod mechanism
 * Support ForkJoinPool trace
 
 #### Documentation